### 2019-07-07 #1548 @whedon
[PRE REVIEW]: infotheory: A C++/Python package for multivariate information theoretic analysis
**Submitting author:** @madvn (<a href="http://orcid.org/0000-0002-9254-8641">Madhavun Candadai</a>)
**Repository:** <a href="https://github.com/madvn/infotheory" target ="_blank">https://github.com/madvn/infotheory</a>
**Version:** 1.0
**Reviewers:** ajgates42, @Autoplectic
