### 2019-07-07 #1548 @labarba
👋 @ madvn — your paper is not compiling; the error appears to be with an ORCID. Can you try to fix, then run `@whedon generate pdf` here to try the compilation again?
