### 2019-07-07 12be72a
updated binning tests
