### 2019-07-07 #1548 @labarba
@madvn — the paper at the moment feels like a "wall of text." Could you make an extra effort to separate the text into sub-sections, and build up to the specialized content a little more gradually? Bear in mind our [documentation](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) says that you should "describing the high-level functionality and purpose of the software for a diverse, non-specialist audience."
