### 2019-07-07 #1548 @usethedata
@labarba I'm a bit strapped at the moment, with ongoing elder care issues and a major proposal due at the end of July.  If there's anyone else that can do it, that might be better.  Otherwise, I can do it, but I'll be slower in getting to things than I'd prefer.
