### 2019-07-07 #1548 @labarba
@poulson — Might you be able to handle this submission? A bit out of your landscape, but no one editor is perfect for this one and you have scads of C++ experience.
