+++ paper/paper.md
@@ -26,11 +26,20 @@ bibliography: paper.bib
 
 # Summary
 
-Information theory was first introduced by Claude Shannon in his seminal paper "A mathematical theory of communication" as a methodology to develop efficient coding and communication of data across noisy channels [@shannon1948]. Its rise to popularity can be primarily attributed to its ability to be applied in any domain, ranging from Economics to Neuroscience. Information theory enables us to answer questions such as, how can we quantify the difference in uncertainty in one random process versus another? and how much does knowing the value of one random variable, reduce uncertainty about another? Until relatively recent times, information theory had been employed to study n-dimensional systems two variables at a time. However, all natural systems are multivariate and a scientific inquiry into their operation requires understanding how these multiple variables interact. In a multivariate system, bivariate measures such as pairwise mutual information alone are insufficient to capture the polyadic interactions between the different variables [@james2017]. There have been multiple approaches proposed to extend bivariate measures to the multivariate case [@williams2010, @griffith2014, @bertschinger2014, @james2018a]. Here we focus on the approach of Partial Information Decomposition (PID) by [@williams2010]. In PID, interactions between variables in a multivariate system can be studied by decomposing the total information that multiple source variables provide about a target variable into its constituent non-negative components. More specifically, in a trivariate case, the three variables can be separated into one target and two source variables. The total information that the two sources have about the target is given by the bivariate mutual information between the concatenated sources as one variable and the target. Using PID, the dependencies between the sources can be studied by decomposing this total information into the following non-negative components: information that each source uniquely provides about the target, information that they redundantly provide and the synergistic information that is only available when both sources are known. Multivariate analysis allows us to ask more detailed questions such as, what is the amount of information that is uniquely provided about a target random variable by one source and not another? and what is the amount of information that is transferred from one random process X to another Y over and above Y’s own information from its past? These questions enable us to understand the interactions between different components of a complex system, thereby leading us towards an understanding of its operation given just the observed data from the system.
+This paper introduces ``infotheory``: a package written in C++ and usable from Python and C++, for multivariate information theoretic analyses of discrete and continuous data. It is open-source ([https://git.io/infot](https://git.io/infot)) and details on how to install it and use it are available on its [website](http://mcandadai.com/infotheory/). This package allows the user to study the relationship between components of a complex system simply from the data recorded during its operation, using the tools of information theory.
 
-This paper introduces ``infotheory``: a package that implements multivariate information theoretic measures for discrete and continuous data. This package implements widely used measures such as entropy and mutual information [@cover2012], as well as more recent measures that arise from multivariate extensions to information theory.
+# Background
+
+Information theory was first introduced by Claude Shannon in his seminal paper "A mathematical theory of communication" as a methodology to develop efficient coding and communication of data across noisy channels [@shannon1948]. Its rise to popularity can be primarily attributed to its ability to be applied in any domain, ranging from Economics to Neuroscience. Information theory provides a general framework to quantify stochastic properties (uncertainty in the outcome of an experiment) and relationships (mutual information that one variable provides about another) between different variables in a system of interest. It provides tools to measure these quantities in a way that is invariant to the scale of the system and allows comparison across systems.
+
+# Statement of need
+
[…]
