### 2019-07-08 #1548 @madvn
Thanks, @labarba. I have reorganized the paper into sections.
