### 2019-07-15 #1548 @poulson
:wave: Alexander (@ajgates42): Given your expertise in python and information theory, you seem like an ideal reviewer for this submission. Would you have the time and interest to help review this submission? And, would you happen to have any suggestions for other reviewers?

:wave: Fabian (@fsteeg): You have listed expertise in information retrieval, though your languages were given as Java and Scala. Would you feel confident as a reviewer of a C++/Python package on information theory? If so, would you have the time to review?

:wave: Andre (@andremrsantos): You listed experience in python/C++ and in statistics and bioinformatics. Would you feel confident reviewing a submission on information theory? If so, do you have the time and interest in doing so?
