### 2019-07-16 #1548 @andremrsantos
Hi @poulson,
although I would love to contribute. My knowledge of information theory is limited and I wouldn't feel confident on review submission on the subject.
I hope you find your reviewers soon.
