### 2019-07-19 #1548 @ajgates42
Thanks for the invite @poulson.  Looks like a fun project, I would be happy to review it.
