### 2019-07-19 #1548 @madvn
@ajgates42 Thanks! Looking forward to hearing your feedback on the project.

@poulson We are continuing to use this project in our research and we have discovered that with some minor changes the multivariate analyses methods in this package can be extended from bi-variate to tri-variate analyses. This would significantly increase the value of the package because, to our knowledge, there are no published packages for tri-variate analyses. Since we are in the pre-review stage, would it be okay for me to update it real quick (in a day or two)?
