### 2019-07-22 #1548 @poulson
Hi @madvn, the purpose of the review is to improve the quality of the software and documentation, so I think waiting for the update would be okay, given that the review hasn't yet started.

I will reach out to a few more potential reviewers in parallel with those changes if that makes sense to you.
