### 2019-07-22 #1548 @poulson
:wave: Fabian (@fsteeg) -- You have listed expertise in information retrieval, though your languages were given as Java and Scala. Would you feel confident as a reviewer of a C++/Python package on information theory? If so, would you have the time to review?

@ajgates42 Do you happen to have any further reviewers you would suggest? Our list of information theory experts is modest and it seems you are from this community. Any help here would greatly aid the review process -- and therefore the work itself.
