### 2019-07-22 #1548 @ajgates42
I think Ryan James (rgjames@ucdavis.edu, @Autoplectic) would be a great
person to have assess the contributions of this package.

Another good fit would be Georg Martius (georg.martius@web.de
, @georgmartius)

On Mon, Jul 22, 2019 at 11:18 AM Jack Poulson <notifications@github.com>
wrote:

> 👋 Fabian (@fsteeg <https://github.com/fsteeg>) -- You have listed
> expertise in information retrieval, though your languages were given as
> Java and Scala. Would you feel confident as a reviewer of a C++/Python
> package on information theory? If so, would you have the time to review?
>
> @ajgates42 <https://github.com/ajgates42> Do you happen to have any
> further reviewers you would suggest? Our list of information theory experts
> is modest and it seems you are from this community. Any help here would
> greatly aid the review process -- and therefore the work itself.
>
