### 2019-07-22 #1548 @poulson
Hi Ryan (@autoplectic) -- You have been recommended as an expert in information theory; are you comfortable enough in C++ and/or python to confidently review this submission? And do you have the time an interest to do so?

Hi Georg (@georgmartius) -- You have also been recommended as an information theory expert. Same question as for Ryan: do you feel confident enough to review the submission, and do you have the time and interest to do so?
