### 2019-07-22 #1548 @georgmartius
Hi Jack,
an even better qualified person would be Joseph Lizier http://lizier.me/joseph/. He wrote JIDT, a quite comprehensive java package.
I am afraid, I cannot invest the appropriate amount of time into it at the moment.
