### 2019-07-22 #1548 @poulson
Hi Joseph (@jlizier) -- your colleague, Georg Martius, recommended you as an expert in information theory software packages. He mentioned that you worked primarily in Java -- would you feel confident reviewing a submission in C++ and Python? And, more importantly, do you have the time and interest to do so?
