### 2019-07-23 #1548 @jlizier
Hi @poulson - main challenge for me atm is time as I'm already loaded up with reviews over the next month. I could commit to early Sept, but I'm guessing that's a bit too long for you? If not, whilst my c++ is a bit rusty I think it would be ok for this
