### 2019-07-23 #1548 @poulson
Thank you for the fast and candid response, Joseph (@jlizier)! I will try to move forward with finding other reviewers and -- if that turns out to run into substantial delays -- circle back with you in early September.
