### 2019-07-29 #1548 @fsteeg
> Fabian (@fsteeg) -- Would you have the time and interest to provide feedback on this submission?

Thanks for getting in touch @poulson, and sorry for the late reply, I was on vacation. My expertise is neither in information theory, nor C++ or Python, so I'm afraid I won't be able to contribute here.
