### 2019-07-30 #1548 @Autoplectic
Hi all, sorry for the delayed response. Yes, I can review it but it might take a couple weeks.
