### 2019-09-29 #1609 @ajgates42
Hi @madvn et al.,
Overall great package, I just have a few suggestions to improve the usability and functionality.

~ the design of the pipeline seems to focus only on continuous data.  This makes its application to discrete data unnecessarily bulky (e.g. I shouldn't need to define a binning for discrete variables).  Similarity when I already know the probability distribution, it would be nice to be able to pass that directly along. Can you streamline the process for these extra use cases?

~ besides dit, you should contrast your package against the other python information theory packages:
[pyentropy](https://www.frontiersin.org/articles/10.3389/neuro.11.004.2009/full)
[IDTxl](https://joss.theoj.org/papers/10.21105/joss.01081)
[JIDT](http://jlizier.github.io/jidt/) (written in java but has a python wrapper)

~ since one of your main claims is the increased efficiency over other packages, please provide a basic and advanced example that demonstrate the packages' performance.

Also, following the JOSS guidelines:
~ clearly state dependences in the installation instructions (it looks like your tests require numpy)
~ all paper reference need a DOIs
