### 2019-10-02 a33938f
some but not all DOIs
