+++ paper/paper.bib
@@ -7,6 +7,7 @@
   pages={379--423},
   year={1948},
   publisher={Wiley Online Library}
+  doi={10.1002/j.1538-7305.1948.tb01338.x}
 }
 
 @book{cover2012,
@@ -14,6 +15,7 @@
   author={Cover, Thomas M and Thomas, Joy A},
   year={2012},
   publisher={John Wiley \& Sons}
+  doi={10.1002/047174882X}
 }
 
 @article{james2017,
@@ -25,6 +27,7 @@
   pages={531},
   year={2017},
   publisher={Multidisciplinary Digital Publishing Institute}
+  doi={10.3390/e19100531}
 }
 
 @article{williams2010,
@@ -41,6 +44,7 @@
   pages={1024--1040},
   year={1985},
   publisher={JSTOR}
+  doi={10.1214/aos/1176349654}
 }
 
 @incollection{griffith2014,
@@ -50,6 +54,7 @@
   pages={159--190},
   year={2014},
   publisher={Springer}
+  doi={10.1007/978-3-642-53734-9_6}
 }
 
 @article{bertschinger2014,
@@ -61,6 +66,7 @@
   pages={2161--2183},
   year={2014},
   publisher={Multidisciplinary Digital Publishing Institute}
+  doi={10.3390/e16042161}
 }
 
 @article{james2018a,
@@ -72,6 +78,7 @@
   pages={014002},
   year={2018},
   publisher={IOP Publishing}
+  doi={10.1088/1751-8121/aaed53}
 }
 
 @article{schreiber2000,
@@ -83,6 +90,7 @@
   pages={461},
   year={2000},
   publisher={APS}
+  doi={10.1103/PhysRevLett.85.461}
 }
 
 @article{williams2011,
@@ -99,6 +107,7 @@
   pages={3--36},
   year={2014},
   publisher={Springer}
+  doi={10.1007/978-3-642-54474-3_1}
 }
 
 @article{james2018b,
