### 2019-10-02 88f3102
DOIs added (except for arxiv papers)
