+++ paper/paper.bib
@@ -6,7 +6,7 @@
   number={3},
   pages={379--423},
   year={1948},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
   doi={10.1002/j.1538-7305.1948.tb01338.x}
 }
 
@@ -14,7 +14,7 @@
   title={Elements of information theory},
   author={Cover, Thomas M and Thomas, Joy A},
   year={2012},
-  publisher={John Wiley \& Sons}
+  publisher={John Wiley \& Sons},
   doi={10.1002/047174882X}
 }
 
@@ -26,7 +26,7 @@
   number={10},
   pages={531},
   year={2017},
-  publisher={Multidisciplinary Digital Publishing Institute}
+  publisher={Multidisciplinary Digital Publishing Institute},
   doi={10.3390/e19100531}
 }
 
@@ -43,7 +43,7 @@
   journal={The Annals of Statistics},
   pages={1024--1040},
   year={1985},
-  publisher={JSTOR}
+  publisher={JSTOR},
   doi={10.1214/aos/1176349654}
 }
 
@@ -53,7 +53,7 @@
   booktitle={Guided Self-Organization: Inception},
   pages={159--190},
   year={2014},
-  publisher={Springer}
+  publisher={Springer},
   doi={10.1007/978-3-642-53734-9_6}
 }
 
@@ -65,7 +65,7 @@
   number={4},
   pages={2161--2183},
   year={2014},
-  publisher={Multidisciplinary Digital Publishing Institute}
+  publisher={Multidisciplinary Digital Publishing Institute},
   doi={10.3390/e16042161}
 }
 
@@ -77,7 +77,7 @@
   number={1},
   pages={014002},
   year={2018},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
   doi={10.1088/1751-8121/aaed53}
 }
 
@@ -89,7 +89,7 @@
   number={2},
   pages={461},
   year={2000},
-  publisher={APS}
+  publisher={APS},
   doi={10.1103/PhysRevLett.85.461}
 }
 
@@ -106,7 +106,7 @@
   booktitle={Directed information measures in neuroscience},
   pages={3--36},
   year={2014},
-  publisher={Springer}
+  publisher={Springer},
   doi={10.1007/978-3-642-54474-3_1}
 }
 
@@ -117,7 +117,8 @@
   volume={3},
   number={25},
   pages={738},
-  year={2018}
+  year={2018},
+  doi={10.21105/joss.00738}
 }
 
 @article{wollstadt12019,
@@ -127,7 +128,8 @@
   volume={4},
   number={34},
   pages={1081},
-  year={2019}
+  year={2019},
+  doi={10.21105/joss.01081}
 }
 
 @article{izquierdo2015,
@@ -138,7 +140,8 @@
   number={10},
   pages={e0140397},
   year={2015},
-  publisher={Public Library of Science}
+  publisher={Public Library of Science},
+  doi={10.1371/journal.pone.0140397}
 }
 
 @article{beer2015,
@@ -149,5 +152,6 @@
   number={1},
   pages={1--38},
   year={2015},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1111/cogs.12142}
 }
