### 2019-10-22 #1609 @kthyng
Hi @madvn! Looks like one review on your JOSS submission has been started. Will you be able to start addressing comments soon?
