### 2019-10-22 #1609 @ajgates42
You might want to try Dr. Artemy Kolchinsky,
https://santafe.edu/people/profile/artemy-kolchinsky
and on Github: artemyk

On Tue, Oct 22, 2019 at 2:18 PM Jack Poulson <notifications@github.com>
wrote:

> Hi @madvn <https://github.com/madvn> 👋. It seems that Ryan James (
> @Autoplectic <https://github.com/Autoplectic>) is unresponsive and we
> will need to find a new reviewer. I already exhausted our relevant reviewer
> pool to get him and @ajgates42 <https://github.com/ajgates42> (thank you
> AJ!) -- do you have any further suggestions for a replacement reviewer?
>
