### 2019-10-29 #1609 @poulson
Hi Artemy (@artemyk) :wave:.

Would you be willing to review this information theory software project, https://github.com/madvn/infotheory, and its associated publication? You were recommended by @ajgates42 (Alexander Gates).
