### 2019-10-29 #1609 @madvn
Hello, all. Sorry about the delay in responding.

Thanks, @ajgates42 for your reviews. I am preparing our response and will submit them in the coming days.

With regards to the second reviewer, Dr. Artemy Kolchinsky, would be great! Dr. Nicholas Timme (http://www.nicholastimme.com/ ; nmtimme on gitHub) is another option.
