### 2019-10-30 #1609 @poulson
Hi Artemy (@artemyk): thank you for being willing to spend time on this. I will reach out to Nicholas Timme as well and see if they can possibly start a bit sooner than that, but, otherwise, a couple of weeks from now would be a good backup plan.

The JOSS editors have put together a very detailed reviewer guide here:
https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
Please let me know if I can answer any more detailed questions about the process. And, again, thank you.
