### 2019-10-30 #1609 @poulson
Hi Nicholas Timme (@nmtimme) :wave:. Madhavun Candadai recommended you as a reviewer for this submission on an information theory software package. Please let me know if you would be willing and able to use your expertise to improve and review this submission.

A detailed reviewer guide can be found at: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
