### 2019-10-30 #1609 @nmtimme
Hello

Thanks for contacting me, but unfortunately I won't be able to help right
now. I would suggest Joe Lizier in Sydney or Ryan James at UC Irvine. Have
a good day!

~Nick

On Wed, Oct 30, 2019 at 5:10 PM Jack Poulson <notifications@github.com>
wrote:

> Hi Nicholas Timme (@nmtimme <https://github.com/nmtimme>) 👋. Madhavun
> Candadai recommended you as a reviewer for this submission on an
> information theory software package. Please let me know if you would be
> willing and able to use your expertise to improve and review this
> submission.
>
> A detailed reviewer guide can be found at:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
