### 2019-10-30 #1609 @poulson
Thank you @nmtimme -- I had reached out to both of them previously and they were unavailable. I think the best option at the moment is to wait a couple of weeks for Artemy (@artemyk) to be available. Thank you for the suggestions!
