### 2019-11-19 #1609 @kthyng
Hi @artemyk (and @poulson) — is now a good time to move forward on your review?
