### 2019-11-20 #1609 @nmtimme
Hi All,

Just to be clear so no one is waiting on me, I'm not available to provide
this review. Thanks!

~Nick

On Tue, Nov 19, 2019, 3:26 PM Artemy Kolchinsky <notifications@github.com>
wrote:

> @kthyng <https://github.com/kthyng> Please ping me in 1 month.
>
