### 2019-11-21 #1609 @kthyng
To summarize for future us:
I've removed one previous reviewer and replaced with @artemyk who plans to review in 4 weeks.
