### 2019-12-29 #1609 @artemyk
Hello @whedon. I have completed my review. Apologies for the delay.

First, I think this is a very welcome package that could be useful to a lot of different research communities.

I have two comments:

1) I agree with all of the issues raised by @ajgates42 .

2) It seems to me that the paper and documentation omits an important piece of information: how is mutual information actually estimated?  Is it by binning, some other method (there is a brief citation to Scott 1985), or something else? It is known that estimating mutual information from a finite sample is a difficult statistical problem, and that most simple estimators are biased (see e.g. Paninski, "Estimation of Entropy and Mutual Information", Neural Computation, 2003). I think its important that the authors (a) prominently state the statistical estimator of mutual information that is being used, (b) explain (even if briefly) how it works (c) explain the bias properties of this estimator, (d) demonstrate that the code correctly implements the estimator by running it on samples (of various size) from distributions with a known entropy, and joint distributions with known MI, and comparing to numerical results. Ideally, (d) would also be implemented in the automated tests of the package.
