### 2019-12-29 #1609 @madvn
Thank you, @ajgates42 and @artemyk . I will respond to all your comments in the coming days.
