+++ benchmarks/mi_baseline.py
@@ -51,22 +51,22 @@ sub_plt(142, [x, y], mi)
 
 # noisy identical variables
 x = np.random.rand(1000)
-y = x + ((np.random.rand(1000) - 0.5) / 10)
-it = infotheory.InfoTools(2, 0)
-it.set_equal_interval_binning([20] * 2, [np.min(x), np.min(y)], [np.max(x), np.max(y)])
+y = x + (np.random.randn(1000) * 0.02)
+it = infotheory.InfoTools(2, 3)
+it.set_equal_interval_binning([50] * 2, [np.min(x), np.min(y)], [np.max(x), np.max(y)])
 it.add_data(np.vstack([x, y]).T)
-mi = it.mutual_info([0, 1]) / np.log2(20)
+mi = it.mutual_info([0, 1]) / np.log2(50)
 sub_plt(143, [x, y], mi)
 
 # random variables
 x = np.random.rand(1000)
 y = np.random.rand(1000)
-it = infotheory.InfoTools(2, 0)
-it.set_equal_interval_binning([20] * 2, [np.min(x), np.min(y)], [np.max(x), np.max(y)])
+it = infotheory.InfoTools(2, 3)
+it.set_equal_interval_binning([50] * 2, [np.min(x), np.min(y)], [np.max(x), np.max(y)])
 it.add_data(np.vstack([x, y]).T)
-mi = it.mutual_info([0, 1]) / np.log2(20)
+mi = it.mutual_info([0, 1]) / np.log2(50)
 sub_plt(144, [x, y], mi)
 
 plt.tight_layout()
-# plt.savefig('mi_baseline.png')
+plt.savefig("mi_baseline.png")
 plt.show()
