+++ paper/paper.md
@@ -11,14 +11,14 @@ tags:
 authors:
   - name: Madhavun Candadai
     orcid: 0000-0002-9254-8641
-    affiliation: "1, 2" # (Multiple affiliations must be quoted)
+    affiliation: "1, 2"
   - name: Eduardo J. Izquierdo
     orcid: 0000-0002-2008-290X
     affiliation: "1, 2"
 affiliations:
- - name: Program in Cognitive Science, Indiana University, Bloomington, IN, U.S.A.
+ - name: Cognitive Science Program, Indiana University, Bloomington, IN, U.S.A.
    index: 1
- - name: School of Informatics, Computing and Engineering, Indiana University, Bloomington, IN, U.S.A.
+ - name: School of Informatics, Computing, and Engineering, Indiana University, Bloomington, IN, U.S.A.
    index: 2
 date: 03 July 2019
 bibliography: paper.bib
@@ -26,7 +26,7 @@ bibliography: paper.bib
 
 # Summary
 
-This paper introduces ``infotheory``: a package written in C++ and usable from Python and C++, for multivariate information theoretic analyses of discrete and continuous data. It is open-source ([https://git.io/infot](https://git.io/infot)) and details on how to install it and use it are available on its [website](http://mcandadai.com/infotheory/). This package allows the user to study the relationship between components of a complex system simply from the data recorded during its operation, using the tools of information theory.
+This paper introduces ``infotheory``: a package written in C++ and usable from Python and C++, for multivariate information theoretic analyses of discrete and continuous data. It is open-source ([https://git.io/infot](https://git.io/infot)) and details on how to install it and use it are available on its [website](http://mcandadai.com/infotheory/). This package allows the user to study the relationship between components of a complex system simply from the data recorded during its operation, using the tools of information theory. Specifically, this package enables the measurement of entropy, and mutual information, and also allows the user to perform partial information decomposition of mutual information into unique, redundant and synergistic information quantities.  
 
 # Background
 
@@ -44,16 +44,21 @@ Multivariate analysis allows us to ask more detailed questions such as, what is 
 
 ``infotheory`` implements widely used measures such as entropy and mutual information [@cover2012], as well as more recent measures that arise from multivariate extensions to information theory.
 As such, the tool has been designed to be easy to use and is ideal for pedagogical demonstrations of information theory as well as in research.
-Here, we highlight seven key aspects of its implementation that make our package a valuable addition to any information theoretic analyses toolbox along with two existing packages, namely dit [@james2018b] (focuses on discrete variables) and IDTxL [@wollstadt12019] (implements an alternate approach to 3 variable PID).
-First, the package is written in C++. One of the main challenges of multivariate analyses on a large, complex system is the amount of computations involved. Implementation in C++ makes the package efficient.
+Here, we highlight seven key aspects of its implementation that make our package a valuable addition to any information theoretic analyses tools set. First, the package is written in C++. One of the main challenges of multivariate analyses on a large, complex system is the amount of computations involved. Implementation in C++ makes the package efficient.
 Second, the package can be used from either C++ or Python. Python wrapping allows for ease of use, as well as compatibility with other powerful open-source libraries such as numpy.
 Third, the API allows adding the data only once to then perform various analyses across different sub-spaces of the dataset cheaply.
-Fourth, a custom sparse data structure is used to represent the random variables. This allows the package to work easily with high-dimensional data.
+Fourth, a custom sparse data structure is used to represent the random variables. This allows the package to work easily with large amounts of data.
 Fifth, to better estimate the data distribution in case of continuous variables, the package employs a kernel-based density estimation method called 'averaged shifted histograms' because of its beneficial trade-off between computational and statistical efficiency [@scott1985].
[…]
