+++ paper/paper.bib
@@ -7,7 +7,7 @@
   pages={379--423},
   year={1948},
   publisher={Wiley Online Library},
-  doi={10.1002/j.1538-7305.1948.tb01338.x}
+  doi={10.1002/j.1538-7305.1948.tb01338.x},
 }
 
 @book{cover2012,
@@ -15,7 +15,7 @@
   author={Cover, Thomas M and Thomas, Joy A},
   year={2012},
   publisher={John Wiley \& Sons},
-  doi={10.1002/047174882X}
+  doi={10.1002/047174882X},
 }
 
 @article{james2017,
@@ -27,7 +27,7 @@
   pages={531},
   year={2017},
   publisher={Multidisciplinary Digital Publishing Institute},
-  doi={10.3390/e19100531}
+  doi={10.3390/e19100531},
 }
 
 @article{williams2010,
@@ -44,7 +44,7 @@
   pages={1024--1040},
   year={1985},
   publisher={JSTOR},
-  doi={10.1214/aos/1176349654}
+  doi={10.1214/aos/1176349654},
 }
 
 @incollection{griffith2014,
@@ -54,7 +54,7 @@
   pages={159--190},
   year={2014},
   publisher={Springer},
-  doi={10.1007/978-3-642-53734-9_6}
+  doi={10.1007/978-3-642-53734-9_6},
 }
 
 @article{bertschinger2014,
@@ -66,7 +66,7 @@
   pages={2161--2183},
   year={2014},
   publisher={Multidisciplinary Digital Publishing Institute},
-  doi={10.3390/e16042161}
+  doi={10.3390/e16042161},
 }
 
 @article{james2018a,
@@ -78,7 +78,7 @@
   pages={014002},
   year={2018},
   publisher={IOP Publishing},
-  doi={10.1088/1751-8121/aaed53}
+  doi={10.1088/1751-8121/aaed53},
 }
 
 @article{schreiber2000,
@@ -90,7 +90,7 @@
   pages={461},
   year={2000},
   publisher={APS},
-  doi={10.1103/PhysRevLett.85.461}
+  doi={10.1103/PhysRevLett.85.461},
 }
 
 @article{williams2011,
@@ -107,7 +107,7 @@
   pages={3--36},
   year={2014},
   publisher={Springer},
-  doi={10.1007/978-3-642-54474-3_1}
+  doi={10.1007/978-3-642-54474-3_1},
 }
 
 @article{james2018b,
@@ -118,7 +118,7 @@
   number={25},
   pages={738},
   year={2018},
-  doi={10.21105/joss.00738}
+  doi={10.21105/joss.00738},
 }
 
 @article{wollstadt12019,
@@ -129,7 +129,7 @@
   number={34},
   pages={1081},
   year={2019},
-  doi={10.21105/joss.01081}
+  doi={10.21105/joss.01081},
 }
 
 @article{izquierdo2015,
@@ -141,7 +141,7 @@
   pages={e0140397},
   year={2015},
   publisher={Public Library of Science},
-  doi={10.1371/journal.pone.0140397}
+  doi={10.1371/journal.pone.0140397},
 }
 
 @article{beer2015,
@@ -153,5 +153,101 @@
   pages={1--38},
   year={2015},
   publisher={Wiley Online Library},
-  doi={10.1111/cogs.12142}
+  doi={10.1111/cogs.12142},
+}
+
+@article{inceinformation,
+  title={An information-theoretic library for the analysis of neural codes},
+  author={Ince, RAA and Bartolozzi, C and Panzeri, S},
+  doi={10.2417/1200906.1663},
+}
+
+@article{lizier2014jidt,
[…]
