### 2020-01-07 51a5fbb
multiple citation syntax fix
