### 2020-01-07 #1609 @madvn
> Hi @madvn et al.,
> Overall great package, I just have a few suggestions to improve the usability and functionality.
>

Hi @ajgates42,
Thank you for review comments. Please find responses below and the updated pdf version of the paper [here](https://github.com/openjournals/joss-papers/blob/joss.01609/joss.01609/10.21105.joss.01609.pdf).

> ~ the design of the pipeline seems to focus only on continuous data. This makes its application to discrete data unnecessarily bulky (e.g. I shouldn't need to define a binning for discrete variables). Similarity when I already know the probability distribution, it would be nice to be able to pass that directly along. Can you streamline the process for these extra use cases?
>

It is indeed true that the focus of this package is on continuous variables. This was deliberate since there are other existing packages that focus on discrete data (e.g., dit). As a result, using this package for discrete data is only done as if it is a special case of continuous data. Similarly, this package is designed for use with large/complex data whose distribution is not known a priori. We have added [line 58](https://github.com/madvn/infotheory/commit/fd5878807a2e0ec52e3d2abfe22c3bbdd2c6ca6b#diff-e2778e7674f45b806282a9611faa7220R58) in the paper to more clearly state the intended use of this package. As it is, the package needs a complete revamp to include these two additional features. If it's okay, we would like to leave extensions to include them for future releases.

> ~ besides dit, you should contrast your package against the other python information theory packages:
> [pyentropy](https://www.frontiersin.org/articles/10.3389/neuro.11.004.2009/full)
> [IDTxl](https://joss.theoj.org/papers/10.21105/joss.01081)
> [JIDT](http://jlizier.github.io/jidt/) (written in java but has a python wrapper)
>

We have included a new short paragraph in the paper at [line 57](https://github.com/madvn/infotheory/commit/fd5878807a2e0ec52e3d2abfe22c3bbdd2c6ca6b#diff-e2778e7674f45b806282a9611faa7220R57) to contrast our package in relation to others.

> ~ since one of your main claims is the increased efficiency over other packages, please provide a basic and advanced example that demonstrate the packages' performance.
>

There are two contexts in this paper where we have mentioned efficiency of the package.
In the first context, we make a general claim with regards to programming languages. We have noted that although this package can be used from Python it is written in C++ which is known to be faster in comparison to Python.
In the second, we point to the computational efficiency of average shifted histograms based on which we have designed a sparse data-structure for efficiency.
Neither of these are claims about the relative computational speed in relation to other packages.
However, because the data is encoded in a sparse data structure, the package scales efficiently with the number of data points and the dimensionality of the dataset.
To illustrate this, we have added data_scaling.py and dim_scaling.py and their associated [results](https://github.com/madvn/infotheory/tree/master/benchmarks#scaling-efficiency---data_scalingpy-dim_scalingpy) in the benchmarks folder of the repo, showing that once the data points have been added, estimating informational quantities take relatively the same amount of time.

> Also, following the JOSS guidelines:
> ~ clearly state dependences in the installation instructions (it looks like your tests require numpy)

Added sub-section on dependencies to README [here](https://github.com/madvn/infotheory#dependencies)

> ~ all paper reference need a DOIs

Done!
