### 2020-01-07 #1609 @madvn
> First, I think this is a very welcome package that could be useful to a lot of different research communities.

Thank you for your comments. Please find responses below and the updated pdf version of the paper [here](https://github.com/openjournals/joss-papers/blob/joss.01609/joss.01609/10.21105.joss.01609.pdf).

>I have two comments:

> 1.  I agree with all of the issues raised by @ajgates42 .

> 2. It seems to me that the paper and documentation omits an important piece of information: how is mutual information actually estimated? Is it by binning, some other method (there is a brief citation to Scott 1985), or something else? It is known that estimating mutual information from a finite sample is a difficult statistical problem, and that most simple estimators are biased (see e.g. Paninski, "Estimation of Entropy and Mutual Information", Neural Computation, 2003). I think its important that the authors
(a) prominently state the statistical estimator of mutual information that is being used,
(b) explain (even if briefly) how it works
(c) explain the bias properties of this estimator,

We have used Average Shifted Histograms as our estimator. As suggested, we have included a new paragraph ([line 55]()) that briefly describes the estimator. Additionally, we have provided general rules for setting up the estimator and have pointed the readers to resources about density estimation and biases.

>(d) demonstrate that the code correctly implements the estimator by running it on samples (of various size) from distributions with a known entropy, and joint distributions with known MI, and comparing to numerical results.
Ideally, (d) would also be implemented in the automated tests of the package.

Yes, we agree that this is important to demonstrate. The current version of the paper and the repository includes these demos. Specifically, we calculate mutual information on three different samples: mutual information on two coupled uniform distributions, two coupled uniform distributions with noise drawn from a Gaussian with 0 mean and standard deviation 0.02, and two independent uniformly random distributions. The results are as expected and are described [here](https://github.com/madvn/infotheory/blob/master/benchmarks/README.md). The first case with identical random variables is also included as part of the automated tests for the package.
