### 2020-01-15 #1609 @artemyk
@kthyng @madvn
The authors have addressed my  concerns.  Thanks, -a
