### 2020-01-23 #1609 @poulson
Hi @ajgates42 -- thank you for all of your volunteering on this so far.

It seems you still haven't checked the following boxes: functionality, performance, installation, community guidelines, authors statement, statement of need, and references.

Do you have lingering concerns on any of these issues? If so, would you mind articulating them?
