### 2020-02-09 #1609 @ajgates42
@kthyng @madvn
The authors have addressed most of my concerns.  Should be a good package for applications of information theory.
