### 2020-02-17 #1609 @poulson
@madvn -- would you mind fixing the one invalid DOI and then registering this package on zenodo and providing the link?
