+++ paper/paper.bib
@@ -215,7 +215,7 @@
   pages={165--176},
   year={2009},
   publisher={Elsevier},
-  doi={10.1016/j.jhydrol.2008.10.019v},
+  doi={10.1016/j.jhydrol.2008.10.019},
 }
 
 @article{scott1979optimal,
