### 2020-02-17 #1609 @madvn
Hi @poulson ,

I've fixed the invalid DOI and have registered with Zenodo at this doi - https://doi.org/10.5281/zenodo.3671994

Please let me know if there is anything else I need to do. Thanks!
