### 2020-03-08 #1609 @Kevin-Mattheus-Moerman
@poulson can you help move this along :point_up: is this ready for acceptance in JOSS?
