### 2020-03-09 #1609 @danielskatz
I've suggested some minor changes in https://github.com/madvn/infotheory/pull/5 and https://github.com/madvn/infotheory/pull/6
