### 2020-03-09 #1609 @madvn
Awesome! 🎉 🎉

Thank you @poulson and @kthyng for your continuous support through this process! Thank you @ajgates42 and @artemyk for your review comments that helped improve the paper and the package overall, and thank you @danielskatz for the final edits!!
