### 2020-03-09 #1609 @danielskatz
At least for me, the paper itself is not appearing (https://www.theoj.org/joss-papers/joss.01609/10.21105.joss.01609.pdf) - once it does, I will close this issue
