### 2020-03-09 7dd3218
Adding JOSS and Zenodo DOIs
